requirements.txt
# Base
pytz==2021.3
python-slugify==5.0.2
Pillow==8.3.2
psycopg2==2.9.1 --no-binary psycopg2

# Django
django==3.2.8

# Environment
django-environ==0.7.0

# Django REST Framework
djangorestframework==3.12.4
django-filter==21.1

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==21.1.0

# Debugging
ipdb==0.13.9

# Tools
django-extensions==3.1.3

# Testing
pytest==6.2.5
pytest-sugar==0.9.4
pytest-django==4.4.0
factory-boy==3.2.0
Faker==8.14.0
freezegun==1.1.0
